requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
timm>=0.9.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# Database
psycopg2-binary>=2.9.7
sqlalchemy>=2.0.0
pandas>=2.0.0

# Image Processing
albumentations>=1.3.0
imageio>=2.31.0

# Utilities
requests>=2.31.0
tqdm>=4.65.0
python-dotenv>=1.0.0

# Jupyter for experiments
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Google Drive API (for dataset download)
gdown>=4.7.0

# Segment Anything Model
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git

# Optional: For better performance
accelerate>=0.21.0